Python/Inverse_Kinematics/pyproject.toml
[tool.poetry]
name = "inverse-kinematics"
version = "0.1.0"
description = ""
authors = ["Steigner"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.23.4"
dill = "^0.3.6"
sympy = "^1.11.1"
pyzmq = "^24.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
